requirements.txt
# # ----------------------------
# # Optional utilities
# # ----------------------------
# schedule==1.2.1

# # retrieval + RAG pipeline + safe refinement
# gpt4all==1.0.3
# scikit-learn==1.3.0

# Core ML & Vector Search
torch>=2.2.0
# faiss-cpu==1.7.3
transformers==4.40.0
sentence-transformers==2.6.0

numpy==1.23.5
scipy==1.10.1
pandas==1.5.3
tqdm==4.65.0

# API Framework
uvicorn==0.21.1
fastapi==0.95.2

# File monitoring
watchdog==4.0.1

# File parsing
openpyxl==3.1.5
pdfplumber==0.11.4
PyPDF2==3.0.1
python-docx==1.1.2
reportlab==3.6.12

# Optional utilities
schedule==1.2.1

# retrieval + RAG pipeline + safe refinement
# gpt4all==1.0.3
scikit-learn==1.3.0

accelerate>=0.21.0


# API Framework
uvicorn==0.21.1
fastapi==0.95.2
flask==3.0.3
flask-cors==4.0.0

# To store in our vector database - Postgres
psycopg2-binary
sqlalchemy
pgvector

python-dotenv==1.0.1